Move result_is_empty from pollen to its sole caller, result_publish
result_is_empty decides a produce-side question (is this finalized dir
worth shipping) with exactly one caller, edge26/result_publish.py; nothing
in pollen/ itself used it, unlike content_type_for in the same file, which
transport.py genuinely shares. Split tests accordingly.
